# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Campus Faso - Gestion du LMD",
        "description": "Documentation générée par Swagger",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/filieres": {
            "get": {
                "tags": [
                    "Filières"
                ],
                "summary": "Récupérer la liste de toutes les filières",
                "operationId": "e342a1c225f63ad5ded6b1b2fda7b9fb",
                "responses": {
                    "200": {
                        "description": "Réponse de succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "departement_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "libelle": {
                                                "type": "string",
                                                "example": "Informatique"
                                            },
                                            "abreviation": {
                                                "type": "string",
                                                "example": "INFO"
                                            },
                                            "responsable_id": {
                                                "type": "string",
                                                "example": "12345"
                                            },
                                            "contact": {
                                                "type": "string",
                                                "example": "+226 25 30 70 65"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "info@uo.bf"
                                            },
                                            "statut": {
                                                "type": "string",
                                                "example": "active"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/filieres/{id}": {
            "get": {
                "tags": [
                    "Filières"
                ],
                "summary": "Récupérer une filière par son ID",
                "operationId": "1ac8a53cc39b0ea843ea62b5331c8142",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la filière",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réponse de succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "departement_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "libelle": {
                                                "type": "string",
                                                "example": "Informatique"
                                            },
                                            "abreviation": {
                                                "type": "string",
                                                "example": "INFO"
                                            },
                                            "responsable_id": {
                                                "type": "string",
                                                "example": "12345"
                                            },
                                            "contact": {
                                                "type": "string",
                                                "example": "+226 25 30 70 65"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "info@uo.bf"
                                            },
                                            "statut": {
                                                "type": "string",
                                                "example": "active"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Filière non trouvée"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/filieres/{filiere}/niveaux": {
            "get": {
                "tags": [
                    "Filières"
                ],
                "summary": "Récupérer la liste des niveaux appartenant à une filière",
                "operationId": "c75c57c4ee6fc5bdb6117600a75af2b8",
                "parameters": [
                    {
                        "name": "filiere",
                        "in": "path",
                        "description": "ID de la filière",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réponse de succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "filiere_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "libelle": {
                                                "type": "string",
                                                "example": "Licence 1"
                                            },
                                            "abreviation": {
                                                "type": "string",
                                                "example": "L1"
                                            },
                                            "accessible": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "statut": {
                                                "type": "string",
                                                "example": "active"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Filière non trouvée"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/filieres/{filiere}/niveaux-accessibles": {
            "get": {
                "tags": [
                    "Filières"
                ],
                "summary": "Récupérer la liste des niveaux accessibles appartenant à une filière",
                "operationId": "1629b4944cdf423da43f4f81f64b8d03",
                "parameters": [
                    {
                        "name": "filiere",
                        "in": "path",
                        "description": "ID de la filière",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réponse de succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "filiere_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "libelle": {
                                                "type": "string",
                                                "example": "Licence 1"
                                            },
                                            "abreviation": {
                                                "type": "string",
                                                "example": "L1"
                                            },
                                            "accessible": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "statut": {
                                                "type": "string",
                                                "example": "active"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Filière non trouvée"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/filieres/{filiere}/diplomes": {
            "get": {
                "tags": [
                    "Filières"
                ],
                "summary": "Récupérer la liste des diplômes appartenant à une filière",
                "operationId": "464087ccc2fd36a534ca28b212a7d3b4",
                "parameters": [
                    {
                        "name": "filiere",
                        "in": "path",
                        "description": "ID de la filière",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réponse de succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "libelle": {
                                                "type": "string",
                                                "example": "Baccalauréat"
                                            },
                                            "abreviation": {
                                                "type": "string",
                                                "example": "BAC"
                                            },
                                            "serie": {
                                                "type": "string",
                                                "example": "C"
                                            },
                                            "option": {
                                                "type": "string",
                                                "example": "Mathématiques"
                                            },
                                            "specialite": {
                                                "type": "string",
                                                "example": "Sciences"
                                            },
                                            "statut": {
                                                "type": "string",
                                                "example": "active"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Filière non trouvée"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
...
    },
...
}